Nett: String with be serialized with the 'best' type
Until now string where always serialized as normal strings where each
special char got escaped. This is OK if only standard strings and no
windows file path's or multiline string are used. Otherwise the TOML
file often was hard to read.

Now when a string is added to a TOML table it gets classified and the
best human readable type is assigned to the string.